made many updates:
Started the day by implementing two custom exceptions: the first handles valid password generation for an account, and the second handles overdraft.

Decided to implement transaction time tracking using the \texttt{java.sql.Timestamp} that is built into the driver as well as \texttt{Timestamp(2)} as the database type.

For security,I'm planning to maintain a map of \texttt{customerIDs} and the hash codes for passwords that can be streamed to a file.

Started implementing the Console Gui